requirements.txt
Flask
pymongo
python-dotenv
bcrypt
flask-wtf